Fix errors on wasm32-unknown-unknown
